# contracts/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "CapabilitiesResponse": {
        "additionalProperties": false,
        "properties": {
          "automatic_sending": {
            "const": false,
            "default": false,
            "title": "Automatic Sending",
            "type": "boolean"
          },
          "deployment_profile": {
            "title": "Deployment Profile",
            "type": "string"
          },
          "knowledge_transport": {
            "const": "mcp",
            "default": "mcp",
            "title": "Knowledge Transport",
            "type": "string"
          },
          "observe_mode": {
            "title": "Observe Mode",
            "type": "boolean"
          },
          "pilot_stage": {
            "enum": [
              "OBSERVE",
              "DRAFT"
            ],
            "title": "Pilot Stage",
            "type": "string"
          },
          "routes": {
            "items": {
              "enum": [
                "AUTO_RESOLVE",
                "CLARIFY",
                "ESCALATE"
              ],
              "type": "string"
            },
            "title": "Routes",
            "type": "array"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "deployment_profile",
          "pilot_stage",
          "observe_mode",
          "routes"
        ],
        "title": "CapabilitiesResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KBMatch": {
        "properties": {
          "article_id": {
            "title": "Article Id",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "retrieval": {
            "additionalProperties": true,
            "title": "Retrieval",
            "type": "object"
          },
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "article_id",
          "title",
          "score",
          "excerpt",
          "category"
        ],
        "title": "KBMatch",
        "type": "object"
      },
...
      "Ticket": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "route": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "ticket_id": {
            "title": "Ticket Id",
            "type": "string"
          },
          "urgency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urgency"
          }
        },
        "required": [
          "ticket_id",
          "subject",
          "body"
        ],
        "title": "Ticket",
        "type": "object"
      },
      "TraceEvent": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "status": {
            "enum": [
              "start",
              "success",
              "warning",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "node",
          "timestamp",
          "duration_ms",
          "status",
          "message"
        ],
        "title": "TraceEvent",
        "type": "object"
      },
      "TriageEnvelope": {
        "additionalProperties": false,
        "properties": {
          "organization_id": {
            "maxLength": 80,
            "minLength": 3,
            "title": "Organization Id",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Request Id",
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/TriageResult"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "organization_id",
          "request_id",
          "result"
        ],
        "title": "TriageEnvelope",
        "type": "object"
      },
      "TriageRequest": {
        "additionalProperties": false,
        "properties": {
          "ticket": {
            "$ref": "#/components/schemas/Ticket"
          }
        },
        "required": [
          "ticket"
        ],
        "title": "TriageRequest",
        "type": "object"
      },
      "TriageResult": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "title": "Citations",
            "type": "array"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "correlation_id": {
            "default": "",
            "title": "Correlation Id",
            "type": "string"
          },
          "decision_summary": {
            "title": "Decision Summary",
            "type": "string"
          },
          "draft": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft"
          },
          "grounding_details": {
            "additionalProperties": true,
            "title": "Grounding Details",
            "type": "object"
          },
          "grounding_validated": {
            "default": false,
            "title": "Grounding Validated",
            "type": "boolean"
          },
          "kb_match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KBMatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "mcp_connected": {
            "default": false,
            "title": "Mcp Connected",
            "type": "boolean"
          },
          "processing_ms": {
            "default": 0,
            "title": "Processing Ms",
            "type": "integer"
          },
          "queue": {
            "default": "General Support",
            "title": "Queue",
            "type": "string"
          },
          "route": {
            "enum": [
              "AUTO_RESOLVE",
              "CLARIFY",
              "ESCALATE"
            ],
            "title": "Route",
            "type": "string"
          },
          "rule_codes": {
            "items": {
              "type": "string"
            },
            "title": "Rule Codes",
            "type": "array"
          },
          "ticket_id": {
            "title": "Ticket Id",
            "type": "string"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/TraceEvent"
            },
            "title": "Trace",
            "type": "array"
          },
          "urgency": {
            "title": "Urgency",
            "type": "string"
          }
        },
        "required": [
          "ticket_id",
          "route",
          "category",
          "urgency",
          "confidence",
          "decision_summary"
        ],
        "title": "TriageResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Private service contract for grounded support-ticket triage.",
    "title": "ResolveFlow AI Internal API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "operationId": "capabilities_v1_capabilities_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-membership-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Membership-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Role"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capabilities",
        "tags": [
          "platform"
        ]
      }
    },
...
    "/v1/triage": {
      "post": {
        "operationId": "triage_v1_triage_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-membership-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Membership-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Role"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Triage",
        "tags": [
          "triage"
        ]
      }
    }
  }
}